+++ content/status.php
@@ -1,19 +1,5 @@
 <article <?php hybrid_attr( 'post' ); ?>>
 
-	<?php
-		$css_class = 'comments-link zero-comments';
-		$number    = (int) get_comments_number( get_the_ID() );
-		$comm_sep  = '<span class="comments-sep no-comments"></span>';
-
-		if ( 1 === $number ) {
-			$css_class = 'comments-link one-comment';
-			$comm_sep  = '<span class="comments-sep"></span>';
-		} elseif ( 1 < $number ) {
-			$css_class = 'comments-link multiple-comments';
-			$comm_sep  = '<span class="comments-sep"></span>';
-		}
-	?>
-
 	<?php if ( is_singular( get_post_type() ) ) : // If viewing a single post. ?>
 
 		<?php if ( get_option( 'show_avatars' ) ) : // If avatars are enabled. ?>
@@ -34,21 +20,7 @@
 			<time <?php hybrid_attr( 'entry-published' ); ?>><?php echo get_the_date(); ?></time>
 			<span class="post-by"><?php echo esc_html_x( 'by', 'post author', 'themelia' ) ?></span>
 			<span <?php hybrid_attr( 'entry-author' ); ?>><?php the_author_posts_link(); ?></span>
-			<?php
-				echo $comm_sep;
-				comments_popup_link(
-					'',
-					sprintf( __( '<span class="comments-number">1</span><span class="comments-txt"> comment</span><span class="screen-reader-text"> on "%1$s"</span>', 'themelia' ),
-						get_the_title()
-					),
-					sprintf( __( '<span class="comments-number">%1$s</span><span class="comments-txt"> comments</span><span class="screen-reader-text"> on "%2$s"</span>', 'themelia' ),
-						$number,
-						get_the_title()
-					),
-					$css_class,
-					''
-				);
-			?>
+			<?php themelia_comments_link(); ?>
 			<?php themelia_edit_link(); ?>
 		</div><!-- .entry-byline -->
 
@@ -78,21 +50,7 @@
 				<time <?php hybrid_attr( 'entry-published' ); ?>><?php echo get_the_date(); ?></time>
 				<span class="post-by"><?php echo esc_html_x( 'by', 'post author', 'themelia' ) ?></span>
 				<a class="entry-permalink" href="<?php the_permalink(); ?>" rel="bookmark" itemprop="url"><?php _e( 'Permalink', 'themelia' ); ?></a>
-				<?php
-					echo $comm_sep;
-					comments_popup_link(
-						'',
-						sprintf( __( '<span class="comments-number">1</span><span class="comments-txt"> comment</span><span class="screen-reader-text"> on "%1$s"</span>', 'themelia' ),
-							get_the_title()
-						),
-					    sprintf( __( '<span class="comments-number">%1$s</span><span class="comments-txt"> comments</span><span class="screen-reader-text"> on "%2$s"</span>', 'themelia' ),
-							$number,
-							get_the_title()
-						),
-					    $css_class,
-					    ''
-					);
-				?>
+				<?php themelia_comments_link(); ?>
 				<?php themelia_edit_link(); ?>
 			</footer><!-- .entry-footer -->